Get Information about a Data Library

The lib_info function returns a data frame of information about each item in the data library. That information includes the item name, file extension, number of rows, number of columns, size in bytes, and the last modified date.

lib_info(x)

Arguments

x
The data library.

Value

A data frame of information about the library.

See also

Examples

# Create temp directory
tmp <- tempdir()

# Create data library
libname(dat, tmp)

#> # library 'dat': 0 items
#> - attributes: rds not loaded
#> - path: C:\Users\User\AppData\Local\Temp\RtmpegGmJe
#> NULL

# Add data to library
lib_add(dat, trees, rock, beaver1)

#> # library 'dat': 3 items
#> - attributes: rds not loaded
#> - path: C:\Users\User\AppData\Local\Temp\RtmpegGmJe
#> - items:
#>      Name Extension Rows Cols   Size        LastModified
#> 1   trees       rds   31    3 2.7 Kb 2020-12-21 23:36:03
#> 2    rock       rds   48    4 3.3 Kb 2020-12-21 23:36:03
#> 3 beaver1       rds  114    4 5.6 Kb 2020-12-21 23:36:03

# Get library information
info <- lib_info(dat)

# Examine info
info

#>      Name Extension Rows Cols   Size        LastModified
#> 1   trees       rds   31    3 2.7 Kb 2020-12-21 23:36:03
#> 2    rock       rds   48    4 3.3 Kb 2020-12-21 23:36:03
#> 3 beaver1       rds  114    4 5.6 Kb 2020-12-21 23:36:03

#      Name Extension Rows Cols   Size        LastModified
# 1 beaver1       rds  114    4 5.3 Kb 2020-11-05 21:27:57
# 2   rocks       rds   48    4 3.1 Kb 2020-11-05 21:27:56
# 3   trees       rds   31    3 2.4 Kb 2020-11-05 21:27:56

# Clean up
lib_delete(dat)