Get the Path for a Data Library

The lib_path function returns the current path of the the library as a string.

lib_path(x)

Arguments

x

The data library.

Value

The path of the data library as a single string.

See also

Examples

# Create temp directory
tmp <- tempdir()

# Create library
libname(dat, tmp)

# Examine library path
lib_path(dat)
# [1] "C:\Users\User\AppData\Local\Temp\RtmpCSJ6Gc"

# Clean up
lib_delete(dat)