Print a data library

A class-specific instance of the print function for data libraries. The function prints the library in a summary manner. Use verbose = TRUE to print the library as a list.

# S3 method for lib
print(x, ..., verbose = FALSE)

Arguments

x
The library to print.

...
Any follow-on parameters.

verbose
Whether or not to print the library in verbose style. By default, the parameter is FALSE, meaning to print in summary style.

Value

The object, invisibly.

See also

Examples

# Create temp directory
tmp <- tempdir()

# Create data library
libname(dat, tmp)

#> # library 'dat': 0 items
#> - attributes: rds not loaded
#> - path: C:\Users\User\AppData\Local\Temp\RtmpA1Jvr8
#> NULL

# Add data to library
lib_add(dat, iris, ToothGrowth, PlantGrowth)

#> # library 'dat': 3 items
#> - attributes: rds not loaded
#> - path: C:\Users\User\AppData\Local\Temp\RtmpA1Jvr8
#> - items:
#>          Name Extension Rows Cols   Size        LastModified
#> 1        iris       rds  150    5 8.1 Kb 2020-12-20 22:52:31
#> 2 ToothGrowth       rds   60    3 3.6 Kb 2020-12-20 22:52:31
#> 3 PlantGrowth       rds   30    2 2.7 Kb 2020-12-20 22:52:31

# Print library summary 
print(dat)

#> # library 'dat': 3 items
#> - attributes: rds not loaded
#> - path: C:\Users\User\AppData\Local\Temp\RtmpA1Jvr8
#> - items:
#>          Name Extension Rows Cols   Size        LastModified
#> 1        iris       rds  150    5 8.1 Kb 2020-12-20 22:52:31
#> 2 ToothGrowth       rds   60    3 3.6 Kb 2020-12-20 22:52:31
#> 3 PlantGrowth       rds   30    2 2.7 Kb 2020-12-20 22:52:31

# library 'dat': 3 items
# - attributes: not loaded
# - path: C:\Users\User\AppData\Local\Temp\RtmpCSJ6Gc
# - items:
#          Name Extension Rows Cols   Size        LastModified
# 1        iris       rds  150    5 7.8 Kb 2020-11-05 22:26:59
# 2 PlantGrowth       rds   30    2 2.5 Kb 2020-11-05 22:26:59
# 3 ToothGrowth       rds   60    3 3.4 Kb 2020-11-05 22:26:59

# Clean up
lib_delete(dat)